Remove the Node type
The idea behind Node was to decorate sqlparser AST nodes with an ID so
that the ID can be used as a key against which to store metadata during
analysis.

The problem is that the ID is fragile as it can only be assigne dduring
traversal and if there are multiple passes of an AST with mutations in
between then the IDs will no longer line up.

Ideally, sqlparser would assign an ID to all nodes during a parse (as
well as other useful information such as spans etc). Maybe one day we'll
fork it to add that, but it would be major surgery.

So what do we use instead? Well, that's coming in the next commit, so
you get to wait and see :)