Implement provenance analysis
What does that even mean?

Essentially, it means that every Expr node within an AST fragment is
tagged with a Source that describes where the value for that Expr comes
from.

E.G. A table+column, subquery, literal value or placeholder parameter
etc.

An Expr has one Source but Source is a can refer to multiple items.

In addition to tagging Expr nodes, SelectItem nodes are tagged.

SelectItem is the type of an AST node that represents a value in a
SELECT projection.

Still to do:

- handle some more Query sub-types
- handle equivalent of projection but for `INSERT`, `UPDATE` and
  `DELETE`.

Note: this includes 100% coverage of sqlparser Expr node variants bnut
not all Statement variants.