feat: add context param to `Transform::transform`
Prior to this commit the only information about a node available to
`Transformer` implementations was its type.

This made contextual transformations hard to implement. For example,
transforming an `Expr` in a projection differently to an `Expr` in an
`ORDER BY` clause.

The context parameter enables context-sensitive transformations of AST
nodes by including the path from the root of the AST to the parent of
`original_node`.