refactor: remove `Semantic` trait and replace with `AsNodeKey` trait
`AsNodeKey` is the *only* way to contruct a `NodeKey` in a consuming
crate now.

`Semantic` was poorly thought out. It existed to prevent a `NodeKey` from
being created from a `Box<N: Visitable>` (because annotations should go on
the inner type).

It was really easy to forget to deref a `Box<N: Visitable>` when creating
a `NodeKey` - annotations end up on the Box instead of the N. Which is
annoying when you're trying to debug why annotations seem to be missing
later on.

So the impl of `AsNodeKey` for `Box<N: AsNodeKey>` dereferences to the
contained type.

AsNodeKey is implemented for `Option<N: 'static>` - returning a
`NodeKey` proxying the `Option` itself - because it might be `None`.