feat: `AsNodeKey::type_name`
I hit an issue with trying to reference a generic argument in a
`tracing::instrument` macro invocation in a downstream crate.

So, make it possible to get the Rust type name of an AST node in places
where using a getting a type name from a generic argument is not possible.